BUG: Move initialization for logger
Move the addition of the stream handler for the logger to the top of the script, outside of the if function that processes the command-line usage. This ensures that even when nrrdify is used in a different python script (preventing the command-line usage function to run), logging is set up correctly.