ENH: Add interface for post processing functions
Add global variable `post_processing`, if set (to a method that accepts the SimpleITK image and a list of pydicom FileDataset objects (representing the slices the image is based on)), this function is applied when generating an image in DicomVolume. This is applied on the scalar volume level (i.e. the image is always a 3D volume, even if the series itself is 4D.