BUG: Fix SOP Class check
Calling `str()` on the pydicom.uid.UID class sometimes returns the UID name (SOP Class name), however, in other cases it returns the UID itself.
Prevent the check from failing when a UID is returned by forcing the check to compare to the name field (always returning the UID name).