ENH: Store number of slices and temporal positions in a volume
If volume is 3D, number of slices is equal to the total count of slices, and temporal positions is 1.

If volume is 4D, check if the total slice count is equally divisible by the number of boundaries, fail if not so.
Additionally, store the number of boundaries as n_slices (done in `sortSlices()`) and store the number of temporal positions found in `sortSlices4D()`.

Finally check the consistency of boundaries when generating 4D volumes.